COM-1443 - create StepCellTitle component
