+++ src/screens/courses/CourseProfile.jsx
@@ -9,8 +9,8 @@ import { LinearGradient } from 'expo-linear-gradient';
 import Courses from '../../api/courses';
 import { WHITE, GREY } from '../../styles/colors';
 import { MAIN_MARGIN_LEFT, ICON, MARGIN } from '../../styles/metrics';
-import OnSiteCell from '../../components/OnSiteCell';
-import ELearningCell from '../../components/ELearningCell';
+import OnSiteCell from '../../components/steps/OnSiteCell';
+import ELearningCell from '../../components/steps/ELearningCell';
 import { ON_SITE, E_LEARNING } from '../../core/data/constants';
 import commonStyles from '../../styles/common';
 import { FIRA_SANS_BLACK } from '../../styles/fonts';
